move upb_utils to src/core/util
