+++ gRPC-C++.podspec
@@ -1325,10 +1325,10 @@ Pod::Spec.new do |s|
                       'src/core/util/string.h',
                       'src/core/util/time_precise.h',
                       'src/core/util/tmpfile.h',
+                      'src/core/util/upb_utils.h',
                       'src/core/util/useful.h',
                       'src/core/xds/grpc/certificate_provider_store.h',
                       'src/core/xds/grpc/file_watcher_certificate_provider_factory.h',
-                      'src/core/xds/grpc/upb_utils.h',
                       'src/core/xds/grpc/xds_audit_logger_registry.h',
                       'src/core/xds/grpc/xds_bootstrap_grpc.h',
                       'src/core/xds/grpc/xds_certificate_provider.h',
@@ -2600,10 +2600,10 @@ Pod::Spec.new do |s|
                               'src/core/util/string.h',
                               'src/core/util/time_precise.h',
                               'src/core/util/tmpfile.h',
+                              'src/core/util/upb_utils.h',
                               'src/core/util/useful.h',
                               'src/core/xds/grpc/certificate_provider_store.h',
                               'src/core/xds/grpc/file_watcher_certificate_provider_factory.h',
-                              'src/core/xds/grpc/upb_utils.h',
                               'src/core/xds/grpc/xds_audit_logger_registry.h',
                               'src/core/xds/grpc/xds_bootstrap_grpc.h',
                               'src/core/xds/grpc/xds_certificate_provider.h',
