+++ gRPC-Core.podspec
@@ -2064,6 +2064,7 @@ Pod::Spec.new do |s|
                       'src/core/util/time_precise.cc',
                       'src/core/util/time_precise.h',
                       'src/core/util/tmpfile.h',
+                      'src/core/util/upb_utils.h',
                       'src/core/util/useful.h',
                       'src/core/util/windows/cpu.cc',
                       'src/core/util/windows/log.cc',
@@ -2076,7 +2077,6 @@ Pod::Spec.new do |s|
                       'src/core/xds/grpc/certificate_provider_store.h',
                       'src/core/xds/grpc/file_watcher_certificate_provider_factory.cc',
                       'src/core/xds/grpc/file_watcher_certificate_provider_factory.h',
-                      'src/core/xds/grpc/upb_utils.h',
                       'src/core/xds/grpc/xds_audit_logger_registry.cc',
                       'src/core/xds/grpc/xds_audit_logger_registry.h',
                       'src/core/xds/grpc/xds_bootstrap_grpc.cc',
@@ -3372,10 +3372,10 @@ Pod::Spec.new do |s|
                               'src/core/util/string.h',
                               'src/core/util/time_precise.h',
                               'src/core/util/tmpfile.h',
+                              'src/core/util/upb_utils.h',
                               'src/core/util/useful.h',
                               'src/core/xds/grpc/certificate_provider_store.h',
                               'src/core/xds/grpc/file_watcher_certificate_provider_factory.h',
-                              'src/core/xds/grpc/upb_utils.h',
                               'src/core/xds/grpc/xds_audit_logger_registry.h',
                               'src/core/xds/grpc/xds_bootstrap_grpc.h',
                               'src/core/xds/grpc/xds_certificate_provider.h',
