+++ grpc.gemspec
@@ -1951,6 +1951,7 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/util/time_precise.cc )
   s.files += %w( src/core/util/time_precise.h )
   s.files += %w( src/core/util/tmpfile.h )
+  s.files += %w( src/core/util/upb_utils.h )
   s.files += %w( src/core/util/useful.h )
   s.files += %w( src/core/util/windows/cpu.cc )
   s.files += %w( src/core/util/windows/log.cc )
@@ -1963,7 +1964,6 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/xds/grpc/certificate_provider_store.h )
   s.files += %w( src/core/xds/grpc/file_watcher_certificate_provider_factory.cc )
   s.files += %w( src/core/xds/grpc/file_watcher_certificate_provider_factory.h )
-  s.files += %w( src/core/xds/grpc/upb_utils.h )
   s.files += %w( src/core/xds/grpc/xds_audit_logger_registry.cc )
   s.files += %w( src/core/xds/grpc/xds_audit_logger_registry.h )
   s.files += %w( src/core/xds/grpc/xds_bootstrap_grpc.cc )
