+++ src/core/xds/grpc/xds_common_types.cc
@@ -45,7 +45,7 @@
 #include <grpc/support/port_platform.h>
 
 #include "src/core/util/json/json_reader.h"
-#include "src/core/xds/grpc/upb_utils.h"
+#include "src/core/util/upb_utils.h"
 #include "src/core/xds/grpc/xds_bootstrap_grpc.h"
 #include "src/core/xds/xds_client/xds_client.h"
